ASSETS-47371: improves prompt
