+++ src/image-alt-text/opportunityHandler.js
@@ -10,10 +10,9 @@
  * governing permissions and limitations under the License.
  */
 
-import { isNonEmptyArray, getPrompt } from '@adobe/spacecat-shared-utils';
-import { FirefallClient } from '@adobe/spacecat-shared-gpt-client';
+import { isNonEmptyArray } from '@adobe/spacecat-shared-utils';
 import { Audit } from '@adobe/spacecat-shared-data-access';
-import { sleep } from '../support/utils.js';
+import { getImageSuggestions } from './suggestionsEngine.js';
 
 /**
  * Synchronizes existing suggestions with new data
@@ -111,88 +110,28 @@ export default async function opportunityAndSuggestions(auditUrl, auditData, con
     throw new Error(`Failed to create alt-text opportunity for siteId ${auditData.siteId}: ${e.message}`);
   }
 
+  const imageUrls = detectedTags.imagesWithoutAltText.map(
+    (image) => new URL(image.src, auditUrl).toString(),
+  );
+
+  const imageSuggestions = await getImageSuggestions(
+    imageUrls,
+    auditUrl,
+    context,
+  );
+
   const suggestions = detectedTags.imagesWithoutAltText.map((image) => ({
     pageUrl: new URL(image.pageUrl, auditUrl).toString(),
     imageUrl: new URL(image.src, auditUrl).toString(),
+    altText: imageSuggestions.filter((res) => res.image_url === image)?.[0]?.suggestion || '',
   }));
-
-  const firefallClient = FirefallClient.createFrom(context);
-  // const prompt = await getPrompt({}, 'image-alt-text', log);
-  // const prompt = 'You re tasked with identifying suitable text for the alt attribute of images. You are an expert SEO consultant, and your goal is to suggest a description for each image that is helpful for the user.### Rules:1. Follow the industry guidelines for accessibility, https://www.w3.org/WAI/tutorials/images/2. Alt-text should reflect how the image relates to the content. Avoid irrelevant descriptions.3. Use natural language, ensuring you\'re not "stuffing" SEO keys.4. If an image is purely decorative and adds no functional or informational value, use an empty string as the alt text.5. For infographics, describe the key data points and trends.6. Ideal description length is 50-60 characters.7. Dont duplicate text thats adjacent in the document or website.8. End the alt text sentence with a period.9. The alt text should be helpful for the user, not the search engine.10. Consider key elements of why you chose this image, instead of describing every little detail. No need to say image of or picture of. But, do say if its a logo, illustration, painting, or cartoon.11. If you can recognize a person in the image, use their name when known.### Response Format:Your response must be a valid JSON object with the following json structure:[{image_url: \'string\',suggestion: \'string\',ai_rationale: \'string\',confidence_score: \'number\'}]**IMPORTANT:**- Provide only the JSON object. Do not include any additional text, explanation, or formatting.### Task:Given this list list of images, suggest a description for the alt attribute of each image that is helpful for the user.';
-
-  const imagesFromHost = [];
-  const otherImages = [];
-  const unsupportedFormatImages = [];
-
-  suggestions.forEach((s) => {
-    const { imageUrl } = s;
-    const regex = /\.(webp|png|gif|jpeg)(?=\?|$)/i;
-
-    if (!regex.test(imageUrl)) {
-      unsupportedFormatImages.push(imageUrl);
-    } else if (imageUrl.includes(auditUrl)) {
-      imagesFromHost.push(imageUrl);
-    } else {
-      otherImages.push(imageUrl);
-    }
-  });
-
-  log.info('Images from host:', imagesFromHost);
-  log.info('Other images:', otherImages);
-  log.info('Unsupported format images:', unsupportedFormatImages);
-
-  const imageList = imagesFromHost;
-  log.info('About to call Firefall with images', imageList);
-
-  function chunkArray(array, chunkSize) {
-    const chunks = [];
-    for (let i = 0; i < array.length; i += chunkSize) {
-      chunks.push(array.slice(i, i + chunkSize));
-    }
-    return chunks;
-  }
-
-  const batchSize = 10;
-  const imageBatches = chunkArray(imageList, batchSize);
-
-  const batchPromises = imageBatches.map(async (batch, index) => {
-    // Introduce a delay of 1 second (1000 milliseconds) for each batch
-    await sleep(index * 5000);
-
-    const firefallOptions = {
-      imageUrls: batch,
-      model: 'gpt-4-vision',
-    };
-    const prompt = await getPrompt({ images: batch }, 'image-alt-text', log);
-    log.info('and prompt', prompt);
-    try {
-      const response = await firefallClient.fetchChatCompletion(prompt, firefallOptions);
-      log.info('Firefall response for alt-text suggestions', JSON.stringify(response));
-      if (response.choices?.length >= 1 && response.choices[0].finish_reason !== 'stop') {
-        log.error('No final suggestions found for batch');
-      }
-
-      const answer = JSON.parse(response.choices[0].message.content);
-      log.info(`Final suggestion for batch, ${JSON.stringify(answer)}`);
-      return answer;
-    } catch (err) {
-      log.error('Error calling Firefall for alt-text suggestion generation for batch', err);
-      return [];
-    }
-  });
-
-  const results = await Promise.all(batchPromises);
-
-  const finalResults = results.reduce((acc, result) => ([...acc, ...result]), []);
-
-  log.info(`Final Merged Suggestions: ${JSON.stringify(finalResults)}`);
 
   await syncAltTextSuggestions({
     opportunity: altTextOppty,
     newSuggestions: suggestions.map((suggestion) => ({
       opportunityId: altTextOppty.getId(),
       type: 'CONTENT_UPDATE',
-      data: { recommendations: [{ ...suggestion, altText: finalResults.filter((res) => res.image_url === suggestion.imageUrl)?.[0]?.suggestion || '' }] },
+      data: { recommendations: [{ ...suggestion }] },
       rank: 1,
     })),
     log,
